Remove padding in the DC(KE) round
The purpose of the padding was to ensure that the DC-net message of
is long enough to be unpredictable for an attacker who does not know
the corresponding NESK (e.g., because the attacker copied a honest
NEPK). This is not necessary. An attacker who does not know the NESK
will be caught in the blame phase because of an unexpected DCEXP
message already.